+++ src/main/java/com/ticketscale/interfaces/rest/FrontendForwardController.java
@@ -1,0 +1,13 @@
+package com.ticketscale.interfaces.rest;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.RequestMapping;
+
+@Controller
+public class FrontendForwardController {
+
+    @RequestMapping(value = {"/admin/**", "/admin"})
+    public String forward() {
+        return "forward:/admin/index.html";
+    }
+}